### 2025-12-06 #9426 @rkurchin
Hi @cralibe, AEiC for the BCM track here just dropping in to let you know that we've got a bit of a backlog on our track at the moment and so I'll have to waitlist this submission for now until an appropriate editor becomes available.

In the meantime, I'll note that this paper is very long for JOSS – we typically target 1000 words or less. Often, this means that things are in the paper that really better belong in the package documentation, so you can perhaps take a look at things through that lens and see if you can trim down.
