+++ paper/paper.md
@@ -65,143 +65,6 @@ We extended our DNCI implementation to handle edge cases common in simulated com
 ## Validation 
 All re-implementations were validated against their R equivalents (see [documentation](https://cralibe.github.io/MetaCommunityMetrics.jl/Validation/)).
 
-# Examples
-## Using rodent metacommunity data as the sample data
-We demonstrate the package using rodent data from the Portal Project [@ernest2018portal], a long-term Chihuahuan desert ecosystem study. The sample dataset includes 21 species across 24 sites from 2010-2023 (117 sampling events). For the ease of demonstration, we simulated our own set of spatial coordinates and environmental variables (temperature, precipitation) for the sample dataset. These data are used to demonstrate the functionailties of DNCI and niche hypervolume measurements. The data can be loaded with:
-```julia
-julia> load_sample_data()
-```
-The complete output of `load_sample_data()` is available in the [documentation](https://cralibe.github.io/MetaCommunityMetrics.jl/#Example).
-
-## Demonstration using the sample data
-Here, we demonstrate a typical community analysis workflow: (1) assessing community stability using variability metrics (`CV_meta()`), (2) decomposing beta diversity to understand spatial and temporal turnover (`spatial_beta_div()`, `temporal_beta_div()`), (3) inferring assembly processes via DNCI analysis (`create_groups()`, `plot_groups()`, `DNCI_multigroup()`), and (4) quantifying niche overlap patterns (`niche_overlap()`, `average_MVNH_det()`, `average_MVNH_dissimilarity()`).
-
-### Assessing temporal stability
-We first examine how abundance variability differs across organizational levels (species vs. community) and spatial scales (local vs. regional) using `CV_meta()`:
-```julia
-julia> using MetaCommunityMetrics, Pipe, DataFrames, Random
-julia> df = load_sample_data()
-julia> CV_meta(df.Abundance, df.Sampling_date_order, df.plot, 
-         df.Species)
-1×4 DataFrame
- Row  CV_s_l   CV_s_r    CV_c_l    CV_c_r   
-      Float64  Float64   Float64   Float64  
-
-   1  1.48859  0.944937  0.718266  0.580183
-```
-Variability in abundance decreases from local to regional scales (comparing `CV_s_l` to `CV_s_r` and `CV_c_l` to `CV_c_r`) and from species to community levels (comparing `CV_s_l` to `CV_c_l` and `CV_s_r` to `CV_c_r`), indicating greater stability at broader spatial scales and higher organizational levels.
-
-### Decomposing beta diversity
-Next, we quantify compositional turnover and determine whether changes are driven by species replacement or abundance differences (`RichDif` equals to abundance differences when abundance data is used and equals to richness differeces when occurrence data is used) using `spatial_beta_div()` and 
-`temporal_beta_div()`. The beta diversity decompositions in space:
-```julia
-julia> spatial_beta_div(df.Abundance, df.Sampling_date_order, 
-         df.plot, df.Species; quant = true)
-1×3 DataFrame
- Row  spatial_BDtotal  spatial_Repl  spatial_RichDif 
-      Float64          Float64       Float64         
-
-   1         0.264822      0.121882         0.142939
-```
-The beta diversity decompositions in time:
-```julia
-julia> temporal_beta_div(df.Abundance, df.Sampling_date_order, 
-         df.plot, df.Species; quant = true)
-1×3 DataFrame
- Row  temporal_BDtotal  temporal_Repl  temporal_RichDif 
-      Float64           Float64        Float64          
-
-   1          0.311222      0.0995483          0.211674
-```
[…]
