+++ paper/paper.bib
@@ -1,19 +1,22 @@
 %%%%My references
 %%R and its package
-@Manual{R,
+ @Manual{R,
     title = {R: A Language and Environment for Statistical Computing},
     author = {{R Core Team}},
-    organization = {`R` Foundation for Statistical Computing},
+    organization = {R Foundation for Statistical Computing},
     address = {Vienna, Austria},
-    year = {2023},
+    year = {2024},
+    doi = {10.32614/r.manuals},
     url = {https://www.R-project.org/},
-}
+  }
+
 
 @Manual{adespatial,
   title   = {adespatial: Multivariate Multiscale Spatial Analysis},
   author  = {St{\'e}phane Dray and David Bauman and Guillaume Blanchet and Daniel Borcard and Sylvie Clappe and Guillaume Gu{\'e}nard and Thibaut Jombart and Guillaume Larocque and Pierre Legendre and Naima Madi and Helene H. Wagner},
   year    = {2023},
   note    = {R package version 0.3-24},
+  doi     = {10.32614/cran.package.adespatial},
   url     = {https://CRAN.R-project.org/package=adespatial},
 }
 
@@ -23,6 +26,7 @@
   year    = {2024},
   note    = {R package version 2.7-0},
   url     = {https://vegandevs.github.io/vegan/},
+  doi     = {10.32614/CRAN.package.vegan},
 }
 
   @Manual{codyn,
@@ -47,6 +51,7 @@
     author = {Jim Hester and Davis Vaughan},
     year = {2025},
     note = {R package version 1.1.4},
+    doi = {10.32614/cran.package.bench},
     url = {https://CRAN.R-project.org/package=bench},
   }
 
@@ -78,7 +83,8 @@
   journal={Journal of Statistical Software},
   volume={107},
   pages={1--32},
-  year={2023}
+  year={2023},
+  doi={10.18637/jss.v107.i04}
 }
 
 @Manual{Pipe,
@@ -228,7 +234,8 @@
   eprint =	 {1608.04295},
   primaryClass = {cs.PF},
   adsurl =	 {https://ui.adsabs.harvard.edu/abs/2016arXiv160804295C},
-  adsnote =	 {Provided by the SAO/NASA Astrophysics Data System}
+  adsnote =	 {Provided by the SAO/NASA Astrophysics Data System},
+  doi={10.48550/arXiv.1608.04295}
 }
 
 %%Papers
@@ -240,6 +247,7 @@
   number={6},
   pages={e3683},
   year={2022},
+  doi={10.1002/ecy.3683},
   publisher={Wiley Online Library}
 }
 
@@ -251,7 +259,8 @@
   number={9},
   pages={1314--1329},
   year={2020},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/ele.13568}
 }
 
 @article{wang2019stability,
@@ -262,6 +271,7 @@
   number={6},
   pages={1200--1211},
   year={2019},
+  doi={10.1111/ecog.04290},
   publisher={Wiley Online Library}
 }
 
@@ -273,6 +283,7 @@
   number={3},
   pages={370--379},
   year={2021},
+  doi={10.1111/ecog.05356},
   publisher={Wiley Online Library}
 }
 
@@ -293,7 +304,8 @@
   number={1},
   pages={379--406},
   year={2010},
-  publisher={Annual Reviews}
+  publisher={Annual Reviews},
+  doi={10.1146/annurev-ecolsys-102209-144621}
 }
 
 @article{leibold2004metacommunity,
@@ -304,15 +316,22 @@
   number={7},
   pages={601--613},
   year={2004},
+  doi={10.1111/j.1461-0248.2004.00608.x},
   publisher={Wiley Online Library}
 }
 
[…]
