+++ paper/paper.bib
@@ -321,16 +321,17 @@
 }
 
 @book{leibold2018metacommunity,
-url = {https://doi.org/10.1515/9781400889068},
-title = {Metacommunity Ecology},
-author = {Mathew A. Leibold and Jonathan M. Chase},
-publisher = {Princeton University Press},
-address = {Princeton},
-doi = {doi:10.1515/9781400889068},
-isbn = {9781400889068},
-year = {2018},
-lastchecked = {2025-12-13}
+ ISBN = {9780691049168},
+ URL = {http://www.jstor.org/stable/j.ctt1wf4d24},
+ author = {Mathew A. Leibold and Jonathan M. Chase},
+ publisher = {Princeton University Press},
+ title = {Metacommunity Ecology},
+ urldate = {2025-12-13},
+ volume = {59},
+ year = {2018},
+ doi={10.2307/j.ctt1wf4d24}
 }
+
 
 
 @article{gonzalez2020scaling,
