+++ paper/paper.md
@@ -49,7 +49,7 @@ Community ecologists are often interested in infering ecological processes from 
 \normalsize
 
 # Statement of need
-Many valuable `R` [@R] packages have been developed to aid in community analyses by providing ecological metrics that can summarize processes from biodiversity patterns and have been widely adopted in the field, such as `adespatial` [@adespatial], `codyn` [@codyn], and `vegan` [@vegan]. However, conducting community analyses in R is often computationally expensive, especially when working with large datasets, a limitation that becomes critical in workflows requiring repeated metric calculations. These include simulation studies for method validation, bootstrap resampling for uncertainty estimation [@efron1992bootstrap], null model approaches for statistical testing [@gotelli1996null], and the analysis of large biodiversity databases, such as [BioTime](https://doi.org/10.1111/geb.70003), all of which can create significant computational bottlenecks when processing multiple large datasets iteratively. To improve the efficiency of these computationally intensive analyses, `Julia` becomes a logical choice for re-implementations. `Julia` is known for its just-in-time compilation that optimizes code execution to levels comparable with lower-level languages like C or Fortran [@Julia-2017], yet Julia remains relatively uncommon in ecological workflows. 
+Many valuable `R` [@R] packages have been developed to aid in community analyses by providing ecological metrics that can summarize processes from biodiversity patterns and have been widely adopted in the field, such as `adespatial` [@adespatial], `codyn` [@codyn], and `vegan` [@vegan]. However, conducting community analyses in R is often computationally expensive, especially when working with large datasets, a limitation that becomes critical in workflows requiring repeated metric calculations. These include simulation studies for method validation, bootstrap resampling for uncertainty estimation [@efron1992bootstrap], null model approaches for statistical testing [@gotelli1996null], and the analysis of large biodiversity databases, such as [BioTime](https://doi.org/10.1111/geb.70003), all of which can create significant computational bottlenecks when processing multiple large datasets iteratively. To improve the efficiency of these computationally intensive analyses, `Julia` becomes a logical choice for re-implementations. `Julia` is known for its just-in-time compilation that optimizes code execution to levels comparable with lower-level languages like C or Fortran [@Julia-2017], yet `Julia` remains relatively uncommon in ecological workflows. 
 
 # State of the field
 To position `MetaCommunityMetrics.jl` within the `Julia` ecosystem, we conducted a comprehensive review of existing `Julia` packages under the [EcoJulia organization](https://ecojulia.org/), a community-driven effort providing tools for ecological and environmental analysis, and the [PoisotLab](https://github.com/PoisotLab), which develops many `Julia` packages focusing on quantitative and computational ecology. We ensured that our package builds beyond the tools currently existing in `Julia` even when sharing similar ecological concepts. For example, our beta diversity metrics share the concept of beta diversity with `Diversity.jl` [@DiversityJL]. However, our implementation decomposes beta diversity into species replacement and richness differences (or abundance difference when abundance data is used instead of occurrence data) [@baselga2010partitioning; @baselga2012relationship; @baselga2013separating; @legendre2014interpreting], which differentiates our approach from `Diversity.jl`, which emphasizes alpha, beta, and gamma diversity measures at the metacommunity and community levels [@reeve2014partition]. To our knowledge, no existing `Julia` packages implement the specific metrics listed in \autoref{functions}. `MetaCommunityMetrics.jl` addresses this gap by providing the first `Julia` implementations of these metrics, which are essential for research in community ecology but were previously only available through computationally expensive R implementations.
