### 2026-03-02 #9426 @espottesmith
Hey @cralibe,

I'm Evan Spotte-Smith, and I'll be serving as the Topic Editor for this submission. To start things off, could you suggest some possible reviewers for your submission? You could take inspiration from the list of similar JOSS papers above, or you could look at our [reviewer database](https://reviewers.joss.theoj.org/reviewers).
