### 2026-03-03 #9426 @espottesmith
@FrancisBanville @gchure @g-rppl @tpoisot

Can any of y'all contribute your expertise to this JOSS submission and serve as reviewers?
