### 2026-03-09 #9426 @espottesmith
@julienmalard @puruckertom @currocam

Can any of y'all contribute your expertise to this JOSS submission and serve as reviewers?
