### 2026-03-12 #10167 @editorialbot
[REVIEW]: MetaCommunityMetrics.jl: A Julia Package for Spatiotemporal Metacommunity Analysis
**Submitting author:** @cralibe (<a href="http://orcid.org/0000-0002-4080-2951">Yan Yin Jenny Cheung</a>)
**Repository:** https://github.com/cralibe/MetaCommunityMetrics.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
