### 2026-04-20 #10167 @espottesmith
@FrancisBanville, @currocam, @julienmalard

Weekly check-in, dear reviewers.
