### 2026-04-24 #10167 @FrancisBanville
Hi @espottesmith ,

I have completed my review and would recommend minor revisions. I have opened two issues in the software repo with my specific comments on the manuscript and documentation, including those concerning the two unchecked boxes.

I have not used generative AI to complete my review.

Please let me know if you need more information or if I can be of any other help.
