### 2026-06-16 c6d4a46
Remove accidentally committed benchmark files and add to .gitignore
