+++ benchmarks/benchmark_r/parallel_log.txt
@@ -1,7188 +1,0 @@
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
-Permutated rows are still empty: manual override 
-  
-
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
-Permutated rows are still empty: manual override 
- 
[…]
