### 2026-06-16 a732d1e
Remove accidentally committed benchmark files and add to .gitignore
