+++ paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-Community ecologists are often interested in infering ecological processes from observational data. This involves calculating different metrics such as beta diversity.  Calculating these metrics is easy and fast due to the availability of `R` packages. However, community ecology today is growing beyond analysing single datasets. Large biodiversity databases have expanded the sets of spatio-temporal community data, and there is a growing interest in using simulations to study ecological processes. For these applications, `R` packages become slow to use. 
+Community ecologists are often interested in infering ecological processes from observational data. This involves calculating different metrics such as beta diversity.  Calculating these metrics is easy and fast due to the availability of `R` packages. However, community ecology today is growing beyond analysing single datasets. Large biodiversity databases have expanded the sets of spatio-temporal community data, and there is a growing interest in using simulations to study ecological processes. For these applications, Julia can provide performance advantages, but having a rich ecosystem of packages can facilitate the usage of Julia. 
 
 `MetaCommunityMetrics.jl` is a `Julia` package that helps ecologists to analyze large datasets of species observations to understand how community composition varies across space and time. By implementing computationally optimized versions of established software and new complementary functions that are designed for handling large spatiotemporal datasets, this package enables ecologists to efficiently analyze large-scale spatiotemporal community datasets that would be computationally prohibitive with existing tools. We designed intuitive interfaces that would feel familiar to ecologists transitioning from `R`. Our function naming conventions closely match their `R` equivalents, and we maintained parameter naming and ordering consistency where possible. Some of the metrics we have implemented include beta diversity decompositions in space and time, dispersal-niche continuum index, and variability metric, but a full list of metrics is listed in \autoref{functions}. Full documentation and examples, including installation instructions and help files, are available at [https://cralibe.github.io/MetaCommunityMetrics.jl/](https://cralibe.github.io/MetaCommunityMetrics.jl/).
 
