### 2026-06-16 8dfc133
updating the documentation and README
