+++ paper/paper.md
@@ -22,57 +22,53 @@ bibliography: paper.bib
 ---
 
 # Summary
-Community ecologists are often interested in infering ecological processes from observational data. This involves calculating different metrics such as beta diversity.  Calculating these metrics is easy and fast due to the availability of `R` packages. However, community ecology today is growing beyond analysing single datasets. Large biodiversity databases have expanded the sets of spatio-temporal community data, and there is a growing interest in using simulations to study ecological processes. While `R` packages can achieve high performance through tools such as `Rcpp`, this requires reimplementation in `C++`, which presents a barrier for most ecologists. `Julia` offers native high performance while remaining as accessible as `R`, making it a more practical path to efficient computation for large-scale ecological analyses. However, broader adoption of `Julia` in ecology depends on having a rich ecosystem of packages, which this package aims to contribute to.
+Community ecologists are often interested in inferring ecological processes from observational data. This involves calculating different metrics such as beta diversity, which measures the dissimilarity between two communities in terms of species richness or abundance. Calculating these metrics is easy and fast due to the availability of `R` packages, particularly for datasets with a limited number of sites and sampling events. However, community ecology today is growing beyond small-scale, single-dataset analyses. Large biodiversity databases have expanded the sets of spatio-temporal community data, and there is a growing interest in using simulations to study ecological processes. While `R` packages can achieve high performance through tools such as `Rcpp`, this requires reimplementation in `C++`, which presents a barrier for most ecologists. `Julia` offers native high performance while remaining as accessible as `R`, making it a more practical path to efficient computation for large-scale ecological analyses. However, broader adoption of `Julia` in ecology depends on having a rich ecosystem of packages, which this package aims to contribute to.
 
-`MetaCommunityMetrics.jl` is a `Julia` package that helps ecologists to analyze large datasets of species observations to understand how community composition varies across space and time. By implementing computationally optimized versions of established software and new complementary functions that are designed for handling large spatiotemporal datasets, this package enables ecologists to efficiently analyze large-scale spatiotemporal community datasets that would be computationally prohibitive with existing tools. We designed intuitive interfaces that would feel familiar to ecologists transitioning from `R`. Our function naming conventions closely match their `R` equivalents, and we maintained parameter naming and ordering consistency where possible. Some of the metrics we have implemented include beta diversity decompositions in space and time, dispersal-niche continuum index, and variability metric, but a full list of metrics is listed in \autoref{functions}. Full documentation and examples, including installation instructions and help files, are available at [https://cralibe.github.io/MetaCommunityMetrics.jl/](https://cralibe.github.io/MetaCommunityMetrics.jl/).
[…]
