+++ paper/paper.bib
@@ -934,3 +934,27 @@ url="https://doi.org/10.1007/978-1-4612-4380-9_41"
   year={2014},
   doi={10.48550/arXiv.1404.6520}
 }
+
+@article{Dornelas2025,
+author = {Dornelas, Maria and Antão, Laura H. and Bates, Amanda E. and Brambilla, Viviana and Chase, Jonathan M. and Chow, Cher F. Y. and Fontrodona-Eslava, Ada and Magurran, Anne E. and Martins, Inês S. and Moyes, Faye and Sagouis, Alban and Adu-Acheampong, Samuel and Acquah-Lamptey, Daniel and Adam, Dušan and Ajani, Penelope A. and Albaina, Aitor and Almaraz, Pablo and An, Jeongseop and Anderson, Roger Sigismund and Anderson, Madelaine Jean Robertson and Antunes, Alexsander Z. and Arismendi, Ivan and Armbrecht, Linda and Aros-Mardones, Pedro and Ashtamoorthy, Sreejith Kalpuzha and Ayyappan, Narayanan and Badihi, Gal and Bailey, Joseph J. and Baird, Andrew H. and Baird, Mark Edward and Balakrishnan, Sreekumar Vadakkethil and Barão-Nóbrega, José António L. and Barash, Adi and Barbosa, Miguel and Barlow, Jos and Bässler, Claus and Beaumont, Matthieu and Beenaerts, Natalie and Begot, Tiago Octavio and Beiroz, Wallace and Beldade, Ricardo and Bell, David M. and Bellgrove, Alecia and Belmaker, Jonathan and Benedetti-Cecchi, Lisandro and Benkwitt, Cassandra E. and Berkum, Pamela Medina-van and Bestelmeyer, Brandon T. and Betts, Matthew G. and Billah, Maxwell Kelvin and Bjorkman, Anne D. and Błażewicz, Magdalena and Bloch, Christopher P. and Blowes, Shane A. and Bode, Antonio and Bogoni, Juliano A. and Bolger, Thomas and Bonebrake, Timothy C. and Bonsdorff, Erik and Bottarin, Roberta and Brokensha, Luke N. and Brooker, Rob W. and Brooks, Andrew J. and Bruelheide, Helge and Bueno, Thiago Almeida and Laguionie, Claire and Campagnoli, Mariana Lopes and Cant, James and Caramaschi, Erica Pellegrini and Caron, Alexandre and Carroll, Tadhg and Caruso, Tancredi and Carvajal-Quintero, Juan and Castaldelli, Giuseppe and Castañeda-Moya, Edward and Castilho, Pedro V. and Cechin, Sonia Zanini and Chaikin, Shahar and Chandrashekara, Uchangi Manjunatha and Chase, Tory J. and Chen, Chaolun Allen and Cherem, Jorge José and Choi, Sei-Woong and Christensen, Erica M. and Christianini, Alexander V. and Chu, Jackson Wing Four and Coad, Peter and Van Colen, Carl and Comte, Lise and Cooper, Elisabeth J. and Cornelissen, J. Hans C. and Cosson, Eddy and Cotano, Unai and Crevecoeur, Luc and Crow, Shannan Kyle and Cumming, Graeme S. and Daga, Vanessa S. and Damasceno, Gabriella and Daskalova, Gergana N. and Davies, Claire H. and Davis, Robert A. and Day, Frank P. and De-La-Zerda, Sussy and Deacon, Amy Elizabeth and de Castro-Arrazola, Indradatta and Degraer, Steven and Deonarinesingh, Kharran and Diaz-Ricaurte, Juan C. and Dickman, Christopher R. and Dirilgen, Tara and Dolan, Ciaran J
[…]
