### 2026-06-18 f04ebfe
update documentation and CI workflow
