+++ paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-Community ecologists are often interested in inferring ecological processes from observational data. This involves calculating different metrics such as beta diversity, which measures the dissimilarity between two communities in terms of species richness or abundance. Calculating these metrics is easy and fast due to the availability of `R` packages, particularly for datasets with a limited number of sites and sampling events. However, community ecology today is growing beyond small-scale, single-dataset analyses. Large biodiversity databases have expanded the sets of spatio-temporal community data, and there is a growing interest in using simulations to study ecological processes. While `R` packages can achieve high performance through tools such as `Rcpp`, this requires reimplementation in `C++`, which presents a barrier for most ecologists. `Julia` offers native high performance while remaining as accessible as `R`, making it a more practical path to efficient computation for large-scale ecological analyses. However, broader adoption of `Julia` in ecology depends on having a rich ecosystem of packages, which this package aims to contribute to.
+Community ecologists are often interested in inferring ecological processes from observational data. This involves calculating different metrics such as beta diversity, which measures the dissimilarity between two communities in terms of species richness or abundance. Calculating these metrics is easy and fast due to the availability of `R` packages, particularly for datasets with a limited number of sites and sampling events. However, community ecology today is growing beyond small-scale, single-dataset analyses. Large biodiversity databases have expanded the sets of spatiotemporal community data, and there is a growing interest in using simulations to study ecological processes. While `R` packages can achieve high performance through tools such as `Rcpp`, this requires re-implementation in `C++`, which presents a barrier for most ecologists. `Julia` offers native high performance while remaining as accessible as `R`, making it a more practical path to efficient computation for large-scale ecological analyses. However, broader adoption of `Julia` in ecology depends on having a rich ecosystem of packages, which this package aims to contribute to.
 
 `MetaCommunityMetrics.jl` is a `Julia` package that helps ecologists analyze datasets of species observations to understand how community composition varies across space and time. By implementing computationally optimized versions of established software and new complementary functions designed for large spatiotemporal datasets, this package enables ecologists to efficiently perform analyses that would be computationally prohibitive with existing tools. We designed intuitive interfaces that would feel familiar to ecologists transitioning from `R`. Our function naming conventions closely match their `R` equivalents, and we maintained parameter naming and ordering consistency where possible. Some of the metrics we have implemented include beta diversity decompositions in space and time, dispersal-niche continuum index, and variability metrics measuring temporal variability in abundance across spatial scales and organizational levels of a metacommunity. A full list is provided in \autoref{functions}. Full documentation and examples, including installation instructions and help files, are available at [https://cralibe.github.io/MetaCommunityMetrics.jl/](https://cralibe.github.io/MetaCommunityMetrics.jl/).
 
@@ -39,7 +39,7 @@ Community ecologists are often interested in inferring ecological processes from
 | | `DNCI_plot_groups()` | A novel visualization function for DNCI analysis |  Visualizes geographic site groupings for DNCI analysis |
 | | `DNCI_multigroup()` | Re-implementation of `R` script from R package `DNCImper` [@DNCImper]| Identifies whether dispersal or niche processes drive community assembly |
 | Niche overlap index | `niche_overlap()` | Novel implementation | Mean niche overlap between all species pairs based on relative abundance distributions |
-| Occupied patches proportion | `prop_patche()` | Re-implementation  of `R` script from @guzman2022accounting |  Proportion of sites occupied by each species |
[…]
