+++ docs/make.jl
@@ -19,8 +19,7 @@ makedocs(
                 "Hypervolume Functions" => "Hypervolume.md",
                 "Validation" => "Validation.md",
                 "Benchmark Results" => "Benchmarking.md",
-                "Demonstration on rodent metacommunity data" => "Example.md",
-                "Acknowledgment" => "Acknowledgment.md"]
+                "Demonstration on rodent metacommunity data" => "Example.md"]
 )
 
 deploydocs(
