### 2026-06-18 3bdd129
update Ci workflow and documentation
