+++ docs/src/NicheOverlapIndex.md
@@ -5,7 +5,7 @@ CurrentModule = MetaCommunityMetrics
 Niche overlap refers to the extent to which different species use the same resources or occupy similar ecological niches. High niche overlap might indicate intense competition, whereas low overlap suggests niche differentiation, allowing species to coexist by minimizing direct competition.
 
 ## Functionality Overview
-To capture the changes in density-independent abiotic response, also known as niche breadth, this implementation developed based on Pianka's Niche Overlap Index, as suggested by Pianka (1973). The summary statistics of this index include the mean, maximum, and minimum values across all species pairs, providing a comprehensive understanding of niche sharing within the community.
+To capture the changes in density-independent abiotic response, also known as niche breadth, this implementation is developed based on Pianka's Niche Overlap Index, as suggested by Pianka (1973). By returning both a summary (the mean, minimum, and maximum values of the pairwise niche overlap index across all species pairs) and the full distribution of pairwise niche overlap values, this function provides a comprehensive understanding of niche sharing within the community.
 
 Since most ecological datasets lack direct measurements of species-specific resource utilization, we adapted the original niche overlap index using abundance data. Specifically, our implementation assumes all species share the same consumption rate and type I functional response, such that species with higher abundance have proportionally higher resource utilization at any given site-time combination. This assumption allows abundance data to serve as a proxy for resource utilization patterns, enabling calculation of niche overlap index from observed abundance data. Keep in mind that this approach will not be able to reflect the fact that some species may be more efficient in consumption than others. Therefore, this approach works best when comparing species with similar body sizes, metabolic rates, or feeding strategies, as these factors strongly influence the abundance-consumption relationship (Brown et al., 2004;  White et a., 2007; Hudson et al., 2013; Kalinoski et al., 2016; Abrams, 2022).
 
