+++ docs/src/Example.md
@@ -36,7 +36,7 @@ julia> temporal_beta_div(df.Abundance, df.Sampling_date_order,
 
    1          0.311222      0.0995483          0.211674
 ```
-Note that `RichDif` here equals to abundance differences because abundance data is used and `quant = true` (`RichDif` equals to richness differeces when occurrance data is used or when `quant = false`). The differences in species composition across space are driven nearly equally by both replacement and abundance differences as the values are very close to each other. This indicates both differences in the total abundance of species, as well as the replacement of species contributes significantly to beta diversity across sites. On the other hand, the differences in community composition across time is primarily driven by abundance differences as the value of abundance differences is higher than that of replacement, meaning that more species are either lost of gained through time which then contributes to temporal beta diversity. This suggests that most sites are experiencing strong seasonal patterns or disturbance over time. 
+Note that `RichDif` here equals to abundance differences because abundance data is used and `quant = true` (`RichDif` equals to richness differences when occurrence data is used or when `quant = false`). The differences in species composition across space are driven nearly equally by both replacement and abundance differences as the values are very close to each other. This indicates both differences in the total abundance of species, as well as the replacement of species contributes significantly to beta diversity across sites. On the other hand, the differences in community composition across time is primarily driven by abundance differences as the value of abundance differences is higher than that of replacement, meaning that more species are either lost of gained through time which then contributes to temporal beta diversity. This suggests that most sites are experiencing strong seasonal patterns or disturbance over time. 
 
 To further determine the relative contribution of the dispersal and niche processes, we will compute DNCI using `DNCI_multigroup()`. Although this index operates on one time point at a time, we do not have to worry about species that are absent or sites that are empty at a given time point, because `DNCI_multigroup()` will filter out species that are completely absent or present everywhere at a given time point in the data before calculating DNCI, and empty sites are allowed in our implementation. 
 
