+++ .github/workflows/CI.yml
@@ -31,6 +31,8 @@ jobs:
             arch: arm64
           - os: macos-latest
             arch: x64
+          - os: ubuntu-latest
+            arch: arm64
         include:
           # Test DataFrames minimum
           - version: '1.10'
