### 2026-06-18 a550714
update paper and documentation
