+++ paper/paper.md
@@ -68,7 +68,7 @@ Finally, to ensure accuracy, all implementations with equivalent `R` functions w
 
 Beyond computational performance, the development of `MetaCommunityMetrics.jl` was driven in part by the needs of an ongoing empirical application of Approximate Bayesian Computation methods to connect metacommunity theory with observed data from a long-term phytoplankton community monitoring dataset. Specifically, the requirement to efficiently calculate summary statistics for this application directly informed the design and implementation of several functions in the package. The manuscript of this application is under preparation.
 
-![A plot showing the speedup of all benchmarked functions across three datasets: small (5,335 observations, 10% of the full dataset), medium (26,676 observations, 50% of the full dataset), and large (53,352 observations), all derived from the sample dataset available via load_sample_data(). The sample dataset is a subset of rodent abundance data from the Portal Project [@ernest2018portal], a long-term study of a Chihuahuan desert ecosystem, covering 24 sites, 21 species, and 117 sampling events from 2010 to 2023. The small and medium datasets were generated by random subsampling of the sample dataset with a fixed random seed and saved for consistent reuse across benchmark runs. Each function was benchmarked using 100 samples in both `Julia` and `R`. Speedup is calculated as the `R` median execution time divided by the `Julia` median execution time. Median speedup and its 95% bootstrap confidence interval (based on 2000 bootstrap iterations) are shown for each comparison. The grey dashed line indicates where speedup equals 1, meaning both implementations require the same execution time. Beta diversity has two variants: `quant = true` (abundance data) and `quant = false` (occurrence data). When `quant = false`, abundance and presence inputs produce equivalent results, but the abundance input requires an additional internal conversion to binary, which accounts for the difference in speedup between the two inputs. Details can be found in the [benchmark results section](https://cralibe.github.io/MetaCommunityMetrics.jl/Benchmarking/).\label{speedup}](speedup.pdf){width=100%}
+![A plot showing the speedup of all benchmarked functions across three datasets: small (5,335 observations, 10% of the full dataset), medium (26,676 observations, 50% of the full dataset), and large (53,352 observations), all derived from the sample dataset available via load_sample_data(). The sample dataset is a subset of rodent abundance data from the Portal Project [@ernest2018portal], a long-term study of a Chihuahuan desert ecosystem, covering 24 sites, 21 species, and 117 sampling events from 2010 to 2023. The small and medium datasets were generated by random subsampling of the sample dataset with a fixed random seed and saved for consistent reuse across benchmark runs. Each function was benchmarked using 100 samples in both `Julia` and `R`. Speedup is calculated as the `R` median execution time divided by the `Julia` median execution time. Median speedup and its 95% bootstrap confidence interval (based on 2000 bootstrap iterations) are shown for each comparison. The grey dashed line indicates where speedup equals 1, meaning both implementations require the same execution time. Beta diversity has two variants: `quant = true` (abundance data) and `quant = false` (occurrence data). When `quant = false`, abundance and presence inputs produce equivalent results, but the abundance input requires an additional internal conversion to binary, which accounts for the difference in speedup between the two inputs. Details can be found in the [documentation](https://cralibe.github.io/MetaCommunityMetrics.jl/Benchmarking/).\label{speedup}](speedup.pdf){width=100%}
 
 # AI usage disclosure
 We used Claude, a generative AI tool, for proofreading and identifying grammatical errors in the manuscript and software documentation, and for providing optimization suggestions during software development. All writing suggestions were manually reviewed by the authors and selectively accepted. For optimization suggestions during software development, all suggestions were manually reviewed by the authors and selectively accepted, verified to produce the same results as before optimization, and benchmarked to confirm improvements in execution time and/or memory usage.
