+++ paper/paper.md
@@ -64,7 +64,7 @@ Second, we prioritized computational efficiency over comprehensive coverage (com
 Finally, to ensure accuracy, all implementations with equivalent `R` functions were validated using the package's built-in sample data by running both `Julia` and `R` implementations on identical inputs within the same script via `RCall`, and computing the absolute differences between outputs. For all deterministic functions, differences were smaller than 0.0001, which we considered negligible and attributable to floating-point differences between implementations. For the DNCI, which is permutation-based, validation was conducted by running both implementations 100 times with 1000 permutations each and comparing the distributions of results, which showed agreement in ecological interpretation across all group pairs (see [documentation](https://cralibe.github.io/MetaCommunityMetrics.jl/Validation/)).
 
 # Research Impact Statement
-`MetaCommunityMetrics.jl` delivers substantial computational advantages over existing tools. For example, we re-implemented the `beta.div.comp` function from the `R` package `adespatial` [@adespatial], achieving up to $\sim$ 51× faster execution times while maintaining methodological consistency (see \autoref{speedup}). As shown in \autoref{speedup}, performance gains range from $\sim$ 3× to 51× across all re-implemented functions using [three testing datasets](https://cralibe.github.io/MetaCommunityMetrics.jl/Benchmarking/#Datesets-used-for-this-benchmark) (large: 53,352 observations; medium: 26,676 observations; small: 5,325 observations), positioning `MetaCommunityMetrics.jl` as an efficient solution for diverse ecological research needs, from standard biodiversity analyses to computationally intensive simulation studies and large-scale data processing. Memory usage is generally comparable between `Julia` and `R` implementations, with notable reductions in `Julia` for the variability metrics function (e.g., 12.60 MiB vs 60.23 MiB for the large dataset). However, the DNCI implementation requires substantially more memory in `Julia` than in `R` (e.g., 803.55 MiB vs 77.79 MiB for the large dataset). This difference is likely related to the fact that the `R` implementation uses parallelization across multiple cores while our `Julia` implementation currently does not, though direct comparisons should be interpreted with caution given the differences in how each language measures memory allocation (see [documentation](https://cralibe.github.io/MetaCommunityMetrics.jl/Benchmarking/) for details).
[…]
