### 2026-06-21 f906fc6
improved DNCI internal functions
