+++ src/Internal.jl
@@ -231,7 +231,7 @@ function check_condition_and_fix(grouped_df)
     return grouped_df
 end
 
-# A function to generate random binary matrices acoording to the assigned algorithm
+# A function to generate random binary matrices according to the assigned algorithm
 # This function is a translation/adaptation of a function from the R package `vegan`, licensed under GPL-2 or later.
 # Original package and documentation available at: https://cran.r-project.org/web/packages/vegan/index.html
 function permatfull(data::Matrix{Int}, fixedmar::String, Nperm::Int) #only for present/absence data
@@ -269,8 +269,6 @@ function permatfull(data::Matrix{Int}, fixedmar::String, Nperm::Int) #only for p
     return permat
     
 end
-
-
     
 # A function to generate null models acoording to the assigned algorithm
 # This function is a translation/adaptation of a function from the R package `vegan`, licensed under GPL-2 or later.
@@ -333,33 +331,42 @@ end
 # This function is a translation/adaptation of a function from the R package `vegan`, licensed under GPL-2 or later.
 # Original package and documentation available at: https://cran.r-project.org/web/packages/vegan/index.html
 function nullmodel_quasiswap(comm::AbstractMatrix{Int}, times::Int)
-
     n, m = size(comm)
     nullmodels = Vector{Matrix{Int}}(undef, times)
 
     for t in 1:times
-        # Initialize nullmodel with the original matrix to preserve sums.
         nullmodel = copy(comm)
-        
-        # Randomly select pairs for potential swaps multiple times.
-        for _ in 1:(10 * n * m) # Increase iterations for more thorough randomization.
-            # Select two distinct rows and columns at random.
-            rows = randperm(n)[1:2]
-            cols = randperm(m)[1:2]
-            
-            # Extract the 2x2 submatrix.
-            submatrix = nullmodel[rows, cols]
-            
-            # Determine if a quasiswap is possible while maintaining row/col sums.
-            if sum(submatrix) == 2 && (submatrix[1, 1] + submatrix[2, 2] == 2 || submatrix[1, 2] + submatrix[2, 1] == 2)
-                # Perform quasiswap if it leads to a valid configuration.
-                nullmodel[rows, cols] .= 1 .- submatrix
+
+        for _ in 1:(10 * n * m)
+            # FIX: draw 2 distinct random indices directly
+            # instead of shuffle! on full array
+            r1 = rand(1:n)
+            r2 = rand(1:n)
+            while r2 == r1
+                r2 = rand(1:n)
+            end
+
+            c1 = rand(1:m)
+            c2 = rand(1:m)
+            while c2 == c1
+                c2 = rand(1:m)
+            end
+
+            a, b = nullmodel[r1, c1], nullmodel[r1, c2]
+            c, d = nullmodel[r2, c1], nullmodel[r2, c2]
+
+            s = a + b + c + d
+            if s == 2 && (a + d == 2 || b + c == 2)
+                nullmodel[r1, c1] = 1 - a
+                nullmodel[r1, c2] = 1 - b
+                nullmodel[r2, c1] = 1 - c
+                nullmodel[r2, c2] = 1 - d
             end
         end
 
         nullmodels[t] = nullmodel
     end
-    
+
     return nullmodels
 end
 
@@ -397,108 +404,79 @@ end
 # This function is a translation/adaptation of a function from the R package `vegan`, licensed under GPL-2 or later.
 # Original package and documentation available at: https://cran.r-project.org/web/packages/vegan/index.html
 function simper(comm::Matrix, groups::Vector)
-    # Set EPS to square root of machine epsilon
     EPS = sqrt(eps(Float64))
-
-    # Here is different from the orginal simper() in R
[…]
