### 2026-06-22 8302169
Bump julia-actions/setup-julia from 1 to 3
Bumps [julia-actions/setup-julia](https://github.com/julia-actions/setup-julia) from 1 to 3.
- [Release notes](https://github.com/julia-actions/setup-julia/releases)
- [Commits](https://github.com/julia-actions/setup-julia/compare/v1...v3)

---
updated-dependencies:
- dependency-name: julia-actions/setup-julia
  dependency-version: '3'
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
