### 2026-07-07 573d6b8
updating .gitignore and documentation
