### 2026-07-07 156288c
updating paper.md and benchmark results
