+++ paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-Community ecologists are often interested in inferring ecological processes from observational data. This involves calculating different metrics such as beta diversity, which measures the dissimilarity between two communities in terms of species richness or abundance. Calculating these metrics is easy and fast due to the availability of `R` packages, particularly for datasets with a limited number of sites and sampling events. However, community ecology today is growing beyond small-scale, single-dataset analyses. Large biodiversity databases have expanded the sets of spatiotemporal community data, and there is a growing interest in using simulations to study ecological processes. While `R` packages can achieve high performance through tools such as `Rcpp`, this requires re-implementation in `C++`, which presents a barrier for most ecologists. `Julia` offers native high performance while remaining as accessible as `R`, making it a more practical path to efficient computation for large-scale ecological analyses. However, broader adoption of `Julia` in ecology depends on having a rich ecosystem of packages, which this package aims to contribute to.
+Community ecologists are often interested in inferring ecological processes from observational data. This involves calculating different metrics such as beta diversity, which measures the dissimilarity between communities in terms of species composition or abundance. Calculating these metrics is easy and fast due to the availability of `R` packages, particularly for datasets with a limited number of sites and sampling events. However, community ecology today is growing beyond small-scale, single-dataset analyses. Large biodiversity databases have expanded the sets of spatiotemporal community data, and there is a growing interest in using simulations to study ecological processes. While `R` packages can achieve high performance through tools such as `Rcpp`, this requires re-implementation in `C++`, which presents a barrier for most ecologists. `Julia` offers native high performance while remaining as accessible as `R`, making it a more practical path to efficient computation for large-scale ecological analyses. However, broader adoption of `Julia` in ecology depends on having a rich ecosystem of packages, which this package aims to contribute to.
 
 `MetaCommunityMetrics.jl` is a `Julia` package that helps ecologists analyze datasets of species observations to understand how community composition varies across space and time. By implementing computationally optimized versions of established software and new complementary functions designed for large spatiotemporal datasets, this package enables ecologists to efficiently perform analyses that would be computationally prohibitive with existing tools. We designed intuitive interfaces that would feel familiar to ecologists transitioning from `R`. Our function naming conventions closely match their `R` equivalents, and we maintained parameter naming and ordering consistency where possible. Some of the metrics we have implemented include beta diversity decompositions in space and time, dispersal-niche continuum index, and variability metrics measuring temporal variability in abundance across spatial scales and organizational levels of a metacommunity. A full list is provided in \autoref{functions}. Full documentation and examples, including installation instructions and help files, are available at [https://cralibe.github.io/MetaCommunityMetrics.jl/](https://cralibe.github.io/MetaCommunityMetrics.jl/).
 
@@ -64,7 +64,7 @@ Second, we prioritized computational efficiency over comprehensive coverage (com
 Finally, to ensure accuracy, all implementations with equivalent `R` functions were validated using the package's built-in sample data by running both `Julia` and `R` implementations on identical inputs within the same script via `RCall`, and computing the absolute differences between outputs. For all deterministic functions, differences were smaller than 0.0001, which we considered negligible and attributable to floating-point differences between implementations. For the DNCI, which is permutation-based, validation was conducted by running both implementations 100 times with 1000 permutations each and comparing the distributions of results, which showed agreement in ecological interpretation across all group pairs (see [documentation](https://cralibe.github.io/MetaCommunityMetrics.jl/Validation/)).
 
[…]
