+++ docs/src/Benchmarking.md
@@ -40,56 +40,56 @@ All times are in millisecond (ms), and memory is in mebibytes (MiB). All values 
 ### Median Execution Time and Speedup Values
 *Bold text indicates the test cases with maximum and minimum speedup values. 95% confidence interval of the speedup is reported.*
 
-| TestCase                                         | Data Size | `Julia` | `R` | Speedup | Lower CI | Upper CI |
-|--------------------------------------------------|----------|--------------|------------|----------------|----------|----------|
-| Beta Diversity (Abundance, quant=true)           | Large    | 0.1812       | 2.4174     | 13.3375        | 12.0799  | 14.5377  |
-| Beta Diversity (Abundance, quant=true)           | Medium   | 0.1114       | 1.3655     | 12.2583        | 11.4627  | 12.8396  |
-| Beta Diversity (Abundance, quant=true)           | Small    | 0.0595       | 1.1525     | 19.3566        | 18.6617  | 20.9457  |
-| Beta Diversity (Abundance, quant=false)          | Large    | 0.0177       | 0.2096     | 11.8243        | 10.8779  | 12.4595  |
-| Beta Diversity (Abundance, quant=false)          | Medium   | 0.0106       | 0.2756     | 25.9352        | 24.2356  | 27.1469  |
-| Beta Diversity (Abundance, quant=false)          | Small    | 0.0065       | 0.3219     | 49.2100        | 46.2983  | 52.4379  |
-| Beta Diversity (Presence, quant=false)           | Large    | 0.0163       | 0.2101     | 12.9270        | 12.3420  | 13.4511  |
-| Beta Diversity (Presence, quant=false)           | Medium   | 0.0106       | 0.3090     | 29.2013        | 28.0925  | 30.1846  |
-| **Beta Diversity (Presence, quant=false)**           | **Small**    | **0.0065**       | **0.3300**     | **51.0969**        | **48.4746**  | **56.4341**  |
-| Spatial Beta Diversity (Abundance, quant=true)   | Large    | 2.7329       | 9.2670     | 3.3909         | 3.3290   | 3.4431   |
-| Spatial Beta Diversity (Abundance, quant=true)   | Medium   | 2.4153       | 9.2043     | 3.8109         | 3.7377   | 3.8663   |
-| Spatial Beta Diversity (Abundance, quant=true)   | Small    | 2.1866       | 8.8448     | 4.0450         | 3.9465   | 4.0956   |
-| Spatial Beta Diversity (Abundance, quant=false)  | Large    | 2.5410       | 6.9715     | 2.7436         | 2.6972   | 2.7788   |
-| Spatial Beta Diversity (Abundance, quant=false)  | Medium   | 2.1578       | 7.1675     | 3.3217         | 3.2773   | 3.3751   |
-| Spatial Beta Diversity (Abundance, quant=false)  | Small    | 2.0421       | 6.7469     | 3.3039         | 3.2016   | 3.4660   |
-| Spatial Beta Diversity (Presence, quant=false)   | Large    | 2.5390       | 6.9638     | 2.7427         | 2.7186   | 2.7809   |
-| Spatial Beta Diversity (Presence, quant=false)   | Medium   | 2.1149       | 7.2182     | 3.4130         | 3.3462   | 3.4824   |
-| Spatial Beta Diversity (Presence, quant=false)   | Small    | 1.8978       | 6.6972     | 3.5290         | 3.3934   | 3.6422   |
-| Temporal Beta Diversity (Abundance, quant=true)  | Large    | 8.1492       | 53.3016    | 6.5407         | 6.4556   | 6.6366   |
[…]
