+++ paper/paper.md
@@ -57,18 +57,18 @@ To position `MetaCommunityMetrics.jl` within the `Julia` ecosystem, we conducted
 # Software Design
 We built `MetaCommunityMetrics.jl` on three design principles: usability, computational efficiency, and accuracy.
 
-First, we prioritized usability for spatiotemporal data and simulation workflows. We extended single-snapshot metrics (single site and/or time point) to operate on multidimensional datasets (multiple sites and/or time points) via aggregation functions (`spatial_beta_div()`, `temporal_beta_div()`, `average_MVNH_det()`, `average_MVNH_dissimilarity()`), eliminating manual looping. Our DNCI implementation also accepts empty sites and singletons using Zero-adjusted Bray-Curtis dissimilarity, filters uninformative species (completely absent or ubiquitous), and implements status flags for five edge cases rather than throwing errors or returning NAs, supporting simulation workflows where sparse matrices are meaningful. We developed `DNCI_create_groups()` and `DNCI_plot_groups()` to automate site grouping (required by @vilmi2021dispersal but manual in `R`) and visualize grouping results. See [documentation](https://cralibe.github.io/MetaCommunityMetrics.jl/) for details.
+First, we prioritized usability for spatiotemporal data and simulation workflows. We extended single-snapshot metrics (single site and/or time point) to operate on multidimensional datasets (multiple sites and/or time points) via aggregation functions (`spatial_beta_div()`, `temporal_beta_div()`, `average_MVNH_det()`, `average_MVNH_dissimilarity()`), eliminating manual looping. Our DNCI implementation also accepts empty sites and singletons using Zero-adjusted Bray-Curtis dissimilarity, filters uninformative species (completely absent or ubiquitous), and implements status flags for five edge cases rather than throwing errors or returning NAs, supporting simulation workflows where sparse matrices are meaningful. We developed `DNCI_create_groups()` and `DNCI_plot_groups()` to automate site grouping (required by @vilmi2021dispersal but manual in `R`) and visualize grouping results.
 
 Second, we prioritized computational efficiency over comprehensive coverage (compared to the `R` implementations). Our beta diversity focuses on the Podani family approach (a mathematical framework that decomposes pairwise dissimilarity between sites into two distinct ecological components: species replacement and richness differences) rather than all five coefficient families, MVNH_dissimilarity() implements only the Bhattacharyya distance (a measure of dissimilarity between two probability distributions used here to quantify niche dissimilarity between two multivariate normal hypervolumes), and DNCI lacks built-in visualizations. We plan to expand these in future releases. To maximize performance, we use vectorized operations where possible and employ Cholesky decomposition in `MVNH_dissimilarity()` for the Mahalanobis component of the Bhattacharyya distance calculation, replacing the LU decomposition (a matrix factorization method that expresses a matrix as the product of a lower and upper triangular matrix) used in solve() in the original `R` implementation, providing substantial performance gains for positive-definite matrices [@haddad2024cholesky].
 
-Finally, to ensure accuracy, all implementations with equivalent `R` functions were validated using the package's built-in sample data by running both `Julia` and `R` implementations on identical inputs within the same script via `RCall`, and computing the absolute differences between outputs. For all deterministic functions, differences were smaller than 0.0001, which we considered negligible and attributable to floating-point differences between implementations. For the DNCI, which is permutation-based, validation was conducted by running both implementations 100 times with 1000 permutations each and comparing the distributions of results, which showed agreement in ecological interpretation across all group pairs (see [documentation](https://cralibe.github.io/MetaCommunityMetrics.jl/Validation/)).
[…]
