+++ paper/paper.bib
@@ -169,6 +169,7 @@
   author = {Dahua Lin and Simon Byrne and Andreas Noack and Douglas Bates and John Myles White and Simon Kornblith and other contributors},
   year   = {2025},
   note   = {Julia package version 0.34.5},
+  doi    = {10.5281/zenodo.20439913},
   url    = {https://github.com/JuliaStats/StatsBase.jl},
 }
 
@@ -211,7 +212,16 @@
   url    = {https://github.com/JuliaMath/Combinatorics.jl},
 }
 
-@article{Robbe2023, doi = {10.21105/joss.05595}, url = {https://doi.org/10.21105/joss.05595}, year = {2023}, publisher = {The Open Journal}, volume = {8}, number = {89}, pages = {5595}, author = {Pieterjan Robbe}, title = {`GaussianRandomFields.jl`: A Julia Package to Generate and Sample from Gaussian Random Fields}, journal = {Journal of Open Source Software} }
+@article{Robbe2023, 
+doi = {10.21105/joss.05595}, 
+url = {https://doi.org/10.21105/joss.05595}, 
+year = {2023}, publisher = {The Open Journal}, 
+volume = {8}, 
+number = {89}, 
+pages = {5595}, 
+author = {Pieterjan Robbe}, 
+title = {`GaussianRandomFields.jl`: A Julia Package to Generate and Sample from Gaussian Random Fields}, 
+journal = {Journal of Open Source Software} }
 
 
 @Manual{ProgressMeter,
@@ -290,11 +300,18 @@
 @article{pianka1973structure,
   title={The Structure of Lizard Communities},
   author={Pianka, Eric R},
-  journal={Annual review of ecology and systematics},
-  pages={53--74},
-  year={1973},
-  publisher={JSTOR}
+  journal= "Annual Review of Ecology, Evolution, and Systematics",
+  year = "1973",
+  volume = "4",
+  number = "Volume 4, 1973",
+  pages = "53-74",
+  doi = "https://doi.org/10.1146/annurev.es.04.110173.000413",
+  url = "https://www.annualreviews.org/content/journals/10.1146/annurev.es.04.110173.000413",
+  publisher = "Annual Reviews",
+  issn = "1545-2069",
+  type = "Journal Article",
 }
+
 
 @article{beaumont2010approximate,
   title={Approximate Bayesian computation in Evolution and Ecology},
@@ -428,10 +445,13 @@
 @article{jaccard1908nouvelles,
   title={Nouvelles Recherches sur la Distribution Florale},
   author={Jaccard, Paul},
-  journal={Bull. Soc. Vaud. Sci. Nat.},
+  journal={Bulletin de la Soci{\'e}t{\xe9} Vaudoise des Sciences Naturelles},
   volume={44},
+  number={163},
   pages={223--270},
-  year={1908}
+  year={1908},
+  doi={10.5169/seals-268384},
+  url={https://doi.org/10.5169/seals-268384}
 }
 
 @article{ruvzivcka1958anwendung,
@@ -439,7 +459,7 @@
   author={Ru{\v{z}}i{\v{c}}ka, M},
   journal={Biologia, Bratislava},
   volume={13},
-  pages={647},
+  pages={647--661},
   year={1958}
 }
 
@@ -528,8 +548,7 @@
 }
 
 @article{baselga2013separating,
-  title={Separating the Two Components of Abundance-Based Dissimilarity: Balanced Changes in Abundance vs. Abundance Gradient},
-  author={Baselga, Andr{\'e}s},
[…]
