### 2026-07-24 #10167 @cralibe
> So, it looks like now, the department is a separate affiliation. Again, the affiliations need to match exactly, so it should be "Department of Entomology, Cornell University, United States"

I just edited it now, they should match now. Thanks.
