### 2026-08-04 77e843f
Add DOI badge to README
Added DOI badge for JOSS paper in README.
